requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
betterproto==1.2.5
black==22.6.0
click==8.1.3
colorama==0.4.5
Django==4.0.6
djangorestframework==3.13.1
dnspython==2.2.1
grpcio==1.47.0
grpcio-tools==1.47.0
grpclib==0.4.2
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
pymongo==4.1.1
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
stringcase==1.2.0
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
